requirements-test.txt
django-nose==0.1.3
nose
nosexcover
mock
selenium
